+++ hex-literal/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+### Changed
+- Disallow comments inside hex strings ([#816])
+- Migrate to 2021 edition and bump MSRV to 1.57 ([#816])
+- Use CTFE instead of proc macro ([#816])
+
+[#816]: https://github.com/RustCrypto/utils/pull/816
+
 ## 0.3.4 (2021-11-11)
 ### Changed
 - Provide more info in `panic!` messages ([#664])